+++ Tests/EditorLogicTests/EditorLogicTestSupport.h
@@ -65,6 +65,8 @@ void testUpdateSessionContainsApplyFailure();
 void testUpdateSessionKeepsPendingUpdateWhenElevationIsCancelled();
 void testUpdateCoordinatorReportsNoPendingRestart();
 void testUpdateCoordinatorContainsAdapterFailure();
+void testVelopackInstallRootFollowsTheCurrentLeafRule();
+void testElevatedCoordinatorArgumentHasOneSpelling();
 void testTheSkinRosterIsTheOneList();
 void testSkinTokensCarryExplicitMode();
 void testTokenSubstitutionOffersAnAlphaForm();